// frontend/src/app/shared/generated/models/create-analysis-template-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateAnalysisTemplateRequest {
  calculationDefinitions?: (string | null);
  maxTolerance?: (number | string | null);
  minTolerance?: (number | string | null);
  name: string;
  site: string;
  testConfiguration?: (string | null);
  validationRules?: (string | null);
}